+++ insight-fe/src/app/(main)/(protected)/educators/theme-builder/components/ColorPaletteManagement.tsx
